// types/types.ts
export interface TokenData {
    address: string
    name: string
    symbol: string
    decimals: number
    totalSupply: string
    priceUsd?: string
    marketCap?: string
    volume24h?: string
    priceChange24h?: string
    priceChange7d?: string
    ath?: string
    blockchain?: string
    launchDate?: string
    description?: string
  }
  
  